Use long press instead of double click
A click is now triggered on releasing the button, while a long press is triggered when the configured time has passed even if the button is still pressed.

This results in an intuitive user experience: To generate a  long press simply hold down the button until the long press has been triggered.

Also, debouncing seems to be unnecessary (at least with the hardware buttons currently used in testing, this may change when the final buttons are in place).